requirements.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.6
djangorestframework==3.6.4
enum34==1.1.6
gunicorn==19.7.1
idna==2.6
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
requests==2.18.4
six==1.10.0
typecraft-python==0.4.4
urllib3==1.22
